Use sys.executable instead of python
`python` command can not exist.
The official is always something like pythonX.Y where X is the major and
Y the minor version.

Sometimes a symlink can exist but python, or python3 is not always
present and can be different from the current python.

One method is `sys.executable` another (on embedded device) is
to extract the version from os.__file__.